refactor add product test
